requirements.txt
gradio==3.50.2
oci==2.125.3
python-dotenv==1.0.1